LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-row-axis-alignment-sticky-positioned-items-002-expected.txt


FAIL .grid 1 assert_equals: 
<div class="container">
    <div style="width: 30px; float:left; height: 10px;"></div>
    <div class="grid">
        <div class="item1 sticky justifySelfStart" data-offset-x="10" data-offset-y="0"></div>
        <div class="item2 sticky justifySelfCenter" data-offset-x="140" data-offset-y="75"></div>
        <div class="item3 sticky justifySelfEnd" data-offset-x="280" data-offset-y="150"></div>
        <div class="item4"></div>
    </div>
    <div style="width: 30px; float:left; height: 10px;"></div>
    <div class="grid scroll">
        <div class="item1 sticky justifySelfStart" data-offset-x="40" data-offset-y="0"></div>
        <div class="item2 sticky justifySelfCenter" data-offset-x="140" data-offset-y="75"></div>
        <div class="item3 sticky justifySelfEnd" data-offset-x="290" data-offset-y="150"></div>
        <div class="item4"></div>
    </div>
    <div style="width: 2000px; height: 10px;"></div>
</div>
offsetLeft expected 140 but got 100
FAIL .grid 2 assert_equals: 
<div class="container">
    <div style="width: 30px; float:left; height: 10px;"></div>
    <div class="grid">
        <div class="item1 sticky justifySelfStart" data-offset-x="10" data-offset-y="0"></div>
        <div class="item2 sticky justifySelfCenter" data-offset-x="140" data-offset-y="75"></div>
        <div class="item3 sticky justifySelfEnd" data-offset-x="280" data-offset-y="150"></div>
        <div class="item4"></div>
    </div>
    <div style="width: 30px; float:left; height: 10px;"></div>
    <div class="grid scroll">
        <div class="item1 sticky justifySelfStart" data-offset-x="40" data-offset-y="0"></div>
        <div class="item2 sticky justifySelfCenter" data-offset-x="140" data-offset-y="75"></div>
        <div class="item3 sticky justifySelfEnd" data-offset-x="290" data-offset-y="150"></div>
        <div class="item4"></div>
    </div>
    <div style="width: 2000px; height: 10px;"></div>
</div>
offsetLeft expected 140 but got 100
